moved vfs changes over to  virtualfilehandler, unverified
