+++ preview/MsixCore/MsixCoreInstaller/Extractor.hpp
@@ -6,15 +6,13 @@
 
 namespace MsixCoreLib
 {
-/// Extractor extracts the files and registry settings from the package to the file system and registry.
+/// Extractor extracts the filesfrom the package to the file system.
 class Extractor : IPackageHandler
 {
 public:
-    /// Extracts the package to the file system and registry.
+    /// Extracts the package to the file system.
     /// Creates the package's root directory in c:\program files\msixcoreapps\<package>
     /// Extracts the files from the package to the package's root directory
-    /// Copies over the VFS files from the package root directory to the actual file system location
-    /// Devirtualizes the registry keys from the package's registry.dat
     HRESULT ExecuteForAddRequest();
 
     /// Removes all the files, directories written during the add.
@@ -58,45 +56,5 @@ private:
     /// Creates the package root directory where all the files will be installed to.
     /// This will be in c:\program files\msixcoreapps\<packagefullname>
     HRESULT CreatePackageRoot();
-    
-    /// Copies a VFS file from the package root to its resolved location.
-    /// for example VFS\ProgramFilesx86\Notepadplusplus\notepadplusplus.exe would get copied over 
-    /// to c:\program files (x86)\Notepadplusplus\notepadplusplus.exe 
-    /// The mapping between the VFS path and the resolved location is obtained through FilePathMappings::GetMap
-    ///
-    /// @param nameStr - A filepath of the file in the VFS 
-    HRESULT CopyVfsFileToLocal(std::wstring nameStr);
-
-    /// Determines if a file needs to be copied. 
-    /// If a file already exists in the target location, the highest version file will be retained
-    /// This follows MSI versioning rules.
-    ///
-    /// @param sourceFullPath - The source file 
-    /// @param targetFullPath - The target location to copy to
-    /// @param needToCopyFile - whether we need to copy the file
-    HRESULT NeedToCopyFile(std::wstring sourceFullPath, std::wstring targetFullPath, bool & needToCopyFile);
-
-    /// Determine whether we need to copy a VFS file from the package root to its resolved location
-    ///
-    /// @param sourceFullPath - The source file 
-    /// @param targetFullPath - The target location to copy to
-    HRESULT CopyVfsFileIfNecessary(std::wstring sourceFullPath, std::wstring targetFullPath);
-
-    /// Removes a VFS file from the resolved location
-    /// This needs to first resolve the VFS file to the real location and then delete it
-    /// It also removes the folder if this is the last file in that folder
-    /// 
-    /// @param fileName - the VFS file name
-    HRESULT RemoveVfsFile(std::wstring fileName);
-
-    /// Resolves the VFS file to the real location
-    ///
-    /// @param fileName - the VFS file name
-    /// @param fileFullPath - the real location full path
-    /// @return E_NOT_SET if the VFS name cannot be found in the mapping.
-    HRESULT ConvertVfsNameToFullPath(std::wstring fileName, std::wstring &fileFullPath);
-
-    /// Removes all VFS files in the package
-    HRESULT RemoveVfsFiles();
 };
 }
